updated code
Updated code for 3 sensors, changed to modifed madgwick filter

+++ benture_box/a_global.ino
@@ -1,10 +1,10 @@
 // manual variables and string read
-double K1, K2, K3;
-double offsetAngle, offsetOmega, winkel_measured;
+float K1, K2, K3;
+float offsetAngle, offsetOmega, winkel_measured;
 String  v1, v2, v3, enable;
 
 // string vars
-double setWrite = 0;
+float setWrite = 0;
 int incomingbyte = 0;
 String stringRead;
 int manualControl;
